Remove function added

+++ sha256.cpp
@@ -135,7 +135,7 @@ unsigned int sign( std::string message, std::string key )
 {
 	unsigned int signedMessage = 0;
 
-	for ( size_t i=0, j=0; i<message.length(); i++, j=(j+1)%key.length() )
+	for ( unsigned int i=0, j=0; i<message.length(); i++, j=(j+1)%key.length() )
 	{
 		unsigned int letter = message[i];
 		unsigned int x_or = key[j];
